server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# EDA and AI Dependencies
pandas==2.2.0
# Pin NumPy to a range compatible with SciPy (SciPy requires NumPy < 2.3.0)
numpy>=1.22.4,<2.3.0
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.12.2
plotly==5.17.0
langgraph==0.2.32
langchain==0.3.5
langchain-core==0.3.13
langchain-groq>=0.1.0
groq>=0.30.0,<1.0
missingno==0.5.2
ydata-profiling==4.6.4
aiofiles==23.2.0

# Dashboard Automation
jinja2==3.1.2
websockets==12.0
jsonschema==4.20.0
markdown==3.5.1
# Database and Authentication
supabase==2.3.3
email-validator==2.1.0

# Additional utilities
pydantic-settings==2.1.0